+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.18.0] - 2024-09-17
+
 ### Added
 
 - The Python version can now be configured using a `.python-version` file. Both the `3.X` and `3.X.Y` version forms are supported. ([#272](https://github.com/heroku/buildpacks-python/pull/272))
@@ -210,7 +212,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation. ([#3](https://github.com/heroku/buildpacks-python/pull/3))
 
-[unreleased]: https://github.com/heroku/buildpacks-python/compare/v0.17.1...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-python/compare/v0.18.0...HEAD
+[0.18.0]: https://github.com/heroku/buildpacks-python/compare/v0.17.1...v0.18.0
 [0.17.1]: https://github.com/heroku/buildpacks-python/compare/v0.17.0...v0.17.1
 [0.17.0]: https://github.com/heroku/buildpacks-python/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/heroku/buildpacks-python/compare/v0.15.0...v0.16.0